Move dev and company iterators into library
